refactor: ScheduleTables 컴포넌트를 새로운 위치로 이동하고, 상태 관리 및 렌더링 최적화
