+++ src/contexts/index.ts
@@ -1,0 +1,6 @@
+export {
+  useSchedulesData,
+  useSchedulesActions,
+  SchedulesProvider,
+  ScheduleProvider,
+} from "./ScheduleContext";
